% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_genes}
\alias{annotate_genes}
\title{annotate the \code{snps} with genes}
\usage{
annotate_genes(snps, all_genes)
}
\arguments{
\item{snps}{a \code{tibble} with at least chrom and pos columns
the chrom column is expected in \verb{1,2,...X,Y} format}

\item{all_genes}{a \code{GRanges} object with the genes, named with \code{ENTREZID}}
}
\value{
the snps \code{tibble} extended with annotated genes
}
\description{
annotate the \code{snps} with genes
}
