% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_queries.R
\name{get_rsid}
\alias{get_rsid}
\title{obtain the rs ids in a range}
\usage{
get_rsid(
  chrom,
  start,
  end,
  snp_mart,
  attributes = c("refsnp_id", "allele", "chrom_start", "chrom_strand")
)
}
\arguments{
\item{chrom}{chromosome in numeric format}

\item{start}{starting position of the range}

\item{end}{ending position of the range}

\item{snp_mart}{a snp biomart obtained with \code{biomaRt::useEnsembl}}

\item{attributes}{a vector with the attributes to obtain. By default we get
"refsnp_id", "allele", "chrom_start", and "chrom_strand"}
}
\value{
a data.frame with the results of the query
}
\description{
obtain the rs ids in a range
}
