% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_funs.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Manhattan plot for the GWAS data}
\usage{
manhattan_plot(
  gwas_data,
  sig = 5e-08,
  sig_color = "red",
  chr_colors = c("grey40", "darkgrey"),
  label_data = NULL,
  label_var = NULL
)
}
\arguments{
\item{gwas_data}{a \code{data.frame} with at least the column chrom, pos, and
p.value}

\item{sig}{a numeric value with the significance level}

\item{sig_color}{color for the significance threshold line}

\item{chr_colors}{colors to use per chromosome, if there are two colors is
going to color odd even chromosomes}

\item{label_data}{a \code{data.frame} with at least the columns chrom, pos, and
\code{label_var}}

\item{label_var}{the variable used to label the data}
}
\value{
a \code{ggplot} object with the Manhattan plot
}
\description{
Manhattan plot for the GWAS data
}
