% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_loci.R
\name{partition_by_snp}
\alias{partition_by_snp}
\title{Separates a \code{tibble} with the summary statistics of a GWAS experiment with
columns id, pos, and p (snp id, snp position, and p.value)}
\usage{
partition_by_snp(data, snp_dist = 250000, pval_thr = 1e-08)
}
\arguments{
\item{data}{A \code{tibble} with at least id, pos and p columns}

\item{snp_dist}{an integer to find all SNPs within that distance from the
significant snp. By default 250 kbps.}

\item{pval_thr}{p.value significant threshold. By default 1e-8.}
}
\value{
a nested \code{tibble} with the separated loci
}
\description{
Separates a \code{tibble} with the summary statistics of a GWAS experiment with
columns id, pos, and p (snp id, snp position, and p.value)
}
