% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_funs.R
\name{qqplot_pvalue}
\alias{qqplot_pvalue}
\title{generate a qqplot for GWAS studies}
\usage{
qqplot_pvalue(gwas_data)
}
\arguments{
\item{gwas_data}{a \code{data.frame} with at least the column chrom, pos, and
p.value}
}
\description{
generate a qqplot for GWAS studies
}
